requirements.txt
numpy==1.20.3
opencv_python==4.5.5.64
PyYAML==6.0